Add more documentation, move linkerdContext struct to separate file
